pyproject.toml
[tool.poetry]
name = "digester"
version = "0.1.0"
description = ""
authors = ["Alex Meyer <acomeyer@gmail.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = {extras = ["async"], version = "^2.3.1"}
gunicorn = "^20.1.0"
openai = "^0.27.4"
tiktoken = "^0.3.3"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.1"
tenacity = "^8.2.2"
flask-cors = "^3.0.10"
docx2txt = "^0.8"
pypdf2 = "^3.0.1"
python-pptx = "^0.6.21"
python-multipart = "^0.0.6"
arrow = "^1.2.3"
pydantic = "^1.10.7"
sqlalchemy = "^2.0.11"
psycopg2 = "^2.9.6"
flask-sqlalchemy = "^3.0.3"
supabase = "^1.0.3"

[tool.poetry.scripts]
start = "server.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
